kill DataServices
