+++ src/Auth/AuthService.ts
@@ -1,5 +1,4 @@
-import { fetchAccount } from '../api';
-import { fetchToken } from '../DataServices/DataServices';
+import { fetchAccount, fetchToken } from '../api';
 import { Form, OtpInfo, Token, User } from './types';
 
 export default class AuthService {
