+++ src/index.tsx
@@ -8,7 +8,6 @@ export * from './ColourScales/ColourScale';
 export * from './ColourScales/ColourScaleHelper';
 export * from './ColourScales/types';
 export * from './common/Metadata/Metadata';
-export * from './DataServices/DataServices';
 export * from './Jobs/JobList/JobList';
 export * from './Jobs/JobList/types';
 export * from './Logs/LogList/LogList';
